+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -150,7 +150,6 @@ func (backend *MetricsReportingKurtosisBackend) CreateAPIContainer(
 	ipAddr net.IP,
 	grpcPortNum uint16,
 	grpcProxyPortNum uint16,
-	enclaveDataDirpathOnHostMachine string,
 	enclaveDataVolumeDirpath string,
 	envVars map[string]string,
 ) (*api_container.APIContainer, error) {
@@ -161,7 +160,6 @@ func (backend *MetricsReportingKurtosisBackend) CreateAPIContainer(
 		ipAddr,
 		grpcPortNum,
 		grpcProxyPortNum,
-		enclaveDataDirpathOnHostMachine,
 		enclaveDataVolumeDirpath,
 		envVars,
 	)