+++ container-engine-lib/docs/changelog.md
@@ -3,6 +3,13 @@
 * Added `ServiceAccounts`, `Roles`, `RoleBindings`, `ClusterRole`, and `ClusterRoleBindings` create and remove methods to `KurtosisManager`
 * Added `CreateEnclave` functionality to kubernetes backend
 
+### Changes
+* Stopped mounting an enclave data directory on the API container
+
+### Breaking Changes
+* Removed the `enclaveDataDirpathOnHostMachine` parameter from `KurtosisBackend.CreateAPIContainer`
+    * Users no longer need to provide this parameter
+
 # 0.17.0
 ### Features
 * Added `PauseService` and `UnpauseService` to `KurtosisBackend`