+++ container-engine-lib/docs/changelog.md
@@ -1,7 +1,7 @@
 # TBD
 ### Features
-* Added `ServiceAccounts`, `Roles`, `RoleBindings`, `ClusterRole`, and `ClusterRoleBindings` create and remove methods to `KurtosisManager`
-* Added `CreateEnclave` functionality to kubernetes backend
+* Added `ServiceAccounts`, `Roles`, `RoleBindings`, `ClusterRole`, and `ClusterRoleBindings` create and remove methods to `KubernetesManager`
+* Added `CreateEnclave` functionality to Kubernetes backend
 
 ### Changes
 * Stopped mounting an enclave data directory on the API container